[BEAM-244] Add RowMapper and ElementInserter to allow user customize the "parsing" of ResultSet/PCollection element
